+++ back/build.gradle
@@ -61,3 +61,11 @@ dependencies {
 tasks.named('test') {
 	useJUnitPlatform()
 }
+
+jar {
+	enabled = false
+}
+
+bootJar {
+	archiveFileName = 'peekaboo-0.0.1-SNAPSHOT.jar'
+}
