--- back/src/main/java/solutiona/challenge/pickaboo/application/dto/LoginUserResponseDto.java
@@ -1,6 +1,0 @@
-package solutiona.challenge.pickaboo.application.dto;
-
-public record LoginUserResponseDto(
-
-) {
-}
