+++ back/src/main/java/solutiona/challenge/pickaboo/application/dto/oauth/LoginUserResponseDto.java
@@ -1,0 +1,6 @@
+package solutiona.challenge.pickaboo.application.dto.oauth;
+
+public record LoginUserResponseDto(
+
+) {
+}
