+++ back/src/main/java/solutiona/challenge/pickaboo/application/service/UpdateUserLanguageService.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import solutiona.challenge.pickaboo.application.usecase.UpdateUserLanguageUseCase;
+import solutiona.challenge.pickaboo.application.usecase.user.UpdateUserLanguageUseCase;
 import solutiona.challenge.pickaboo.domain.entity.User;
 import solutiona.challenge.pickaboo.domain.repository.UserRepository;
