+++ back/src/main/java/solutiona/challenge/pickaboo/application/service/oauth/LoginByKakaoService.java
@@ -5,9 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import solutiona.challenge.pickaboo.application.dto.JwtTokenDto;
-import solutiona.challenge.pickaboo.application.dto.LoginResponseDto;
-import solutiona.challenge.pickaboo.application.dto.OAuth2UserInfoResponseDto;
+import solutiona.challenge.pickaboo.application.dto.oauth.JwtTokenDto;
+import solutiona.challenge.pickaboo.application.dto.oauth.LoginResponseDto;
 import solutiona.challenge.pickaboo.application.usecase.oauth.LoginByGoogleUseCase;
 import solutiona.challenge.pickaboo.core.util.JwtUtil;
 import solutiona.challenge.pickaboo.core.util.OAuth2Util;
