+++ back/src/main/java/solutiona/challenge/pickaboo/application/usecase/oauth/LoginByGoogleUseCase.java
@@ -1,6 +1,6 @@
 package solutiona.challenge.pickaboo.application.usecase.oauth;
 
-import solutiona.challenge.pickaboo.application.dto.LoginResponseDto;
+import solutiona.challenge.pickaboo.application.dto.oauth.LoginResponseDto;
 import solutiona.challenge.pickaboo.core.annotation.UseCase;
 
 @UseCase
