+++ back/src/main/java/solutiona/challenge/pickaboo/application/usecase/user/ReadLoginUserUsecase.java
@@ -1,8 +1,7 @@
-package solutiona.challenge.pickaboo.application.usecase;
+package solutiona.challenge.pickaboo.application.usecase.user;
 
-import solutiona.challenge.pickaboo.application.dto.JwtTokenDto;
+import solutiona.challenge.pickaboo.application.dto.oauth.JwtTokenDto;
 import solutiona.challenge.pickaboo.core.annotation.UseCase;
-import solutiona.challenge.pickaboo.core.security.JwtDto;
 import solutiona.challenge.pickaboo.presentation.request.LoginUserRequestDto;
 
 @UseCase
